refactor the settings validation code
