+++ splink/linker.py
@@ -83,7 +83,8 @@ from .pipeline import SQLPipeline
 from .predict import predict_from_comparison_vectors_sqls
 from .profile_data import profile_columns
 from .settings import Settings
-from .settings_validator import InvalidSettingsLogger
+from splink.settings_validation.column_lookups import InvalidColumnsLogger
+from splink.settings_validation.valid_types import InvalidTypesAndValuesLogger
 from .splink_comparison_viewer import (
     comparison_viewer_table_sqls,
     render_splink_comparison_viewer_html,
@@ -467,11 +468,13 @@ class Linker:
         ):
             return
 
-        self.settings_validator = InvalidSettingsLogger(self)
-        self.settings_validator._validate_dialect()
+        # Run miscellaneous checks on our settings dictionary.
+        settings_invalid_types_values = InvalidTypesAndValuesLogger(self)
+        settings_invalid_types_values._validate_dialect()
+
         # Constructs output logs for our various settings inputs
         if validate_settings:
-            self.settings_validator.construct_output_logs()
+            InvalidColumnsLogger(self).construct_output_logs()
 
     def _initialise_df_concat(self, materialise=False):
         cache = self._intermediate_table_cache
